requirements.txt
fastapi==0.45.0
uvicorn==0.10.8
requests==2.22.0
pyjwt==1.7.1
cryptography==2.8
python-dotenv==0.10.3
